requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/dev.in
#
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
boto3==1.34.74
    # via -r requirements/base.in
botocore==1.34.74
    # via
    #   boto3
    #   s3transfer
build==1.2.1
    # via pip-tools
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
coverage[toml]==7.5.3
    # via pytest-cov
django==4.2
    # via
    #   -r requirements/base.in
    #   django-debug-toolbar
    #   django-dramatiq
    #   model-bakery
django-debug-toolbar==4.3.0
    # via -r requirements/dev.in
django-dramatiq==0.11.6
    # via -r requirements/base.in
django-environ==0.11.2
    # via -r requirements/base.in
django-ratelimit==4.1.0
    # via -r requirements/base.in
dramatiq[rabbitmq]==1.16.0
    # via
    #   -r requirements/base.in
    #   django-dramatiq
execnet==2.1.1
    # via pytest-xdist
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.22.0
    # via -r requirements/base.in
jsonschema-specifications==2023.12.1
    # via jsonschema
model-bakery==1.18.0
    # via -r requirements/dev.in
packaging==24.0
    # via
    #   build
    #   pytest
pika==1.3.2
    # via dramatiq
pip-tools==7.4.1
    # via -r requirements/dev.in
pluggy==1.5.0
    # via pytest
prometheus-client==0.20.0
    # via dramatiq
psycopg[binary]==3.1.18
    # via -r requirements/base.in
psycopg-binary==3.1.18
    # via psycopg
pymemcache==4.0.0
    # via -r requirements/base.in
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
pytest==8.2.0
    # via
    #   -r requirements/dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-xdist
pytest-cov==5.0.0
    # via -r requirements/dev.in
pytest-django==4.8.0
    # via -r requirements/dev.in
pytest-xdist==3.6.1
    # via -r requirements/dev.in
python-dateutil==2.9.0.post0
    # via botocore
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.2
    # via
    #   -r requirements/base.in
    #   requests-mock
requests-mock==1.12.1
    # via -r requirements/dev.in
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
ruff==0.4.5
    # via -r requirements/dev.in
s3transfer==0.10.1
    # via boto3
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.10.0
    # via psycopg
urllib3==2.2.1
    # via
    #   botocore
    #   requests
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
